Impose bidirectional flow in/out buffer with goal
As pipes have unidirectional flow, the hot/cold pipe of the buffer is respresented by two different ones. If the optimization problem is left free, it will find a solution for which there is non-zero flow in both pipes. (This is a downside of modelling the buffer using simply conservation of heat.)

To find a solution that has physical meaning, we add an extra goal where the
discharge in/out of the buffer is minimized. As this goal is after the
heat demand/supply goals, it does not affect the usage of the buffer per
se.

There is a toggle to activate or not this goal. If the goal is active, a
small slack is used.

Toggle to decide whether to make the plots or not.