Add HeatNetworkMixin and move generic functionality from examples to it
We use the newly added functionality of String parameters/constants to
RTC-Tools 2.5.0a2 (and PymocaV 0.8.0) to detect components in the model.

For pipes we move the functionality of the (inequality) head loss to the
HeatNetworkMixin. There are three options that a user can choose from
with regards to head loss in pipes, see the `HeadLossOption` enum:

1. No head loss constraints
2. C*Q^2
3. Linearized version of Darcy-Weisbach head loss

The user can change the default (option 2) with the `head_loss_option`
key in the `heat_network_options` method.

Note that route pipe lists for plots are not moved, as they are project
specific.