HNMixin: Support generic interface to model components
With Modelica only one of the ways in which a model can be developed
(ESDL being the future other major one), we need to make sure we have a
uniform way of defining what components are in the model regardless of
what model mixin the user is using. The user needs to explicitly
implement this function themselves, or use one  of the provided mixins
to do it for them.