Prepare for possible flow definition change
Instead of the sign of flows being equal (Q_in = Q_out), we can consider
changing it to Q_in = -Q_out. Flow _into_ a component would then be
positive, and flow _out of_ a component would be negative. The changes
in this commit make this possible future change of definition simpler,
but does not change anything yet.