Move to dynamically mixing nodes for bidirectional flow
Some networks have bidirectional flow, and the flow can therefore
reverse from time step to time step. The definition of what is an
inflow/outflow to the node is then no longer constant.

To support this behavior, it is necessary to add mixing constraints to
the system on a time step basis. This has to take into account the
orientation of the pipe (i.e. is the upstream or downstream connector of
the pipe connected to the node) and the flow direction at a certain time
step.

It is not possible to have the flow direction be undetermined in the
QTHMixin, as those models will not converge. In other words, it is
necessary to explicitly set the flow direction in the pipes per time
step. Eventually, figuring out the flow direction is something that can
be done by e.g. a (mixed-integer) linear "heat" matching model.

To summarize, two new pieces of information are necessary, and have to
be provided somehow:

1. Topology of the network (nodes <-> pipes)
2. Flow direction in pipes per time step

The topology of the network contains information about which pipes are
connected to what nodes, which connector of said node, and the
orientation of the connection (upstream or downstream side of the pipe).
Because not all future models will be made in Modelica, a `Topology`
class has been defined that returns this information in a certain way.
All classes that are an instance of `BaseComponentTypeMixin` need to
return such a Topology instance. For Modelica models, the
`ModelicaComponentTypeMixin` has been amended to deduce this topological
information.

As far as the flow direction is concerned, a new abstract property has
been added to QTHMixin. This will map each pipe to a Timeseries in
`constant_inputs`. For models like the basic_buffer example, where the
flow direction of pipes is constant throughout time, we have also added
a convience class that deduces the flow direction from the bounds that
are set in the Modelica model (`BoundsToPipeFlowDirectionsMixin`), and
implements this property such that the user does not have to do it
anymore.