QTHMixin: Improve max temp roc constraints
The overall objective of the constraints is to ensure that in the entire
network the rate of change of the temperature is below a certain
threshold. Imposing temperature derivative constraints on all the pipes
serves this purpose but at the price of:

I) imposing rather complicated constraints;
II) adding redundant and empty constraints.

Note: I) is due to the fact that constraints on the rate of change of
the temperature imply that the Q of the pipe has to be at least a
certain value, which depends on T. Thus the task of minimizing the heat
losses in the network, while satisfying certain heat requests, increases
in complexity substantially.

II) Redundant constraints arises from, e.g. setting constraints on both
supply and return lines at the demand components. Empty constraints are
due to initial derivatives.

To avoid both these isses, we add the constraints:

1) on the states (not on the derivatives). Thus skipping t0.
2) Only on supply temperatures of sources and on bidirectional flow
pipes.

The reasoning behind 2) is that: provided there is enough flow in the
network, i.e. heat losses are not substantial, controlling the
temperatures of the sources is enough provided that there a no
bidirectional flow pipes. However, when bidirectionality happens, if we
want to ensure a rate of change constraint this has to be explicitely
set. This is the default option. If one does not want a rate of change
constraint on bidirectional pipes, set `max_t_der_bidirect_pipe` to
False.

Note: 1) and 2) do not ensure that roc of temperature of all the pipes
within the network is exactly within wanted range. However, it is a good
proxy.