Add HeatMixin for models made with Heat sublibrary
Similiar to the QTHMixin (and its corresponding QTH sublibrary), the
HeatMixin will add helper variables and constraints to the problem.

The HeatMixin logic assumes that there is no heat in the return line.
If the solution allows heat in the return line, we cannot use the
HeatMixin to infer flow directions for the QTHMixin. Note that this is
why some Heat components (e.g., source and demand) have zero heat in the
supply line. Also note that there is a check in post() to ensure that
this assumption holds.