Add 2-stage convenience function 'run_heat_network_optimization'
With the QTH problems needing the flow direction fixed, and the Heat
problem suitable to solve for flow directions, we add a function to
which the user can pass both problems. The function will then run them
in series, and pass the flow direction information from the Heat problem
to the QTH problem.