Modelica: Add Heat sublibrary
The Heat sublibrary is a mixed-integer linear model of transport of
heat. The mixed-integerness follows from allowing flow direction
changes, and the heat loss being dependent on the direction (i.e. the
heat loss is either `up - down` or `down - up`).

The Heat sublibrary can be used stand-alone, but can also serve a
preprocessor to calculate flow directions that can subsequently be used
in a QTH model.

Note that the naming of components and their parameters is deliberately
similar if not equal between the QTH and Heat sublibraries. This is to
aid in the preprocessor-role described above, and in being able to
easily set dynamic nominals/bounds on both models.