QTHMixin: Fix max temp constraints
If the source delivers at constant temperature (for at least one time
step), the creation of the constraint would fail due to a length
mismatch, because we did not slice `dt` with `inds` yet.

Furthermore, we want to avoid making empty constraints, so we skip the
creation of the MX expression in case there is nothing to be done.