Modify buffer component and example
The buffer component is modelled as a two tanks buffer: hot and cold.

The overall volume is constant. Indeed, when the buffer charges, hot
water flows into the hot tank and cold water gets out. The opposite
happens when the buffer discharges. A buffer might also neither charge
nor discharge. To allow for this option, the hot/cold pipes attached to
the buffer must be disableble.

To correctly model temperature mixing, one has to know in advance
whether the buffer is charging, discharging or neither, in a certain
timestep. The heat model can be used to extract such information.

Heat losses are included in this model. To exclude heat
losses, one has to set the heat tranfer coefficient to zero.

The basic example has been changed appropriately.