QTHMixin: Fix empty constraints wit no head loss option
When the user selects the no head loss option, there is no pressure
change in the network. That is, all the pipes have dH equal to zero.

To avoid the creation of trivial constraints, we change the following:
- set bounds of dH to be zero, instead of creating a constraint;
- skip all the other head loss related constraints.

Note that now the H's throughout the network do not need to be the same.
However, the H's of the network do not have a physical meaning whatever
head loss option is chosen.