ModelicaCompType: Fix detection of Node connections
A Heat Pipe has both a property `HeatOut.Heat` and `Heat_out`. The will
both turn up as aliases of the Node's `HeatConn[i]`, triggering the
assertion failure that only one component should be connected.

Instead, we should make sure to also filter the aliases on (one of) the
property of the connector, which for Heat connectors is `.Heat`, and for
QTH connectors is `.T`.