Modelica: Remove superfluous theta parameters from Buffer, Node and Pipe
A `theta` parameter is not needed in the Modelica equations for the
Buffer, the Node and the Pipe. For the Buffer, the examples already did
not set the `theta` parameter equal to the global one. For the Pipe and
Node we remove the corresponding `theta` modifiers from the examples.

Note that the QTHMixin can still use the global `theta` parameter to set
additional constraints that relate to the Node, Pipe and Buffer
components. This commit just clarifies that the _Modelica_ equation
block for each of these three components does not use the `theta`
parameter.