CI/CD: Workaround default GitLab runner failing
The default GitLab runner was upgraded to 13.5.0, but apparently the
docker installation still used the (obsolete) aufs driver. These two are
incompatible, leading to the test stage failing to download the
artifacts from the build stage.

See https://gitlab.com/gitlab-org/gitlab-runner/-/issues/26639 for more
details on this issue.

The workaround of this commit was made possible by adding a new runner
that still uses an older version (13.1.2). We tag our jobs such that
this runner is used, as this runner will not pick up untagged jobs.