Add ESDL support
Aside from models being specified in Modelica or PyCML directly by the
user, we now also support converting ESDL models to a PyCML one. ESDL is
a modelling language created to describe complete (hybrid) energy
systems in one uniform format [0]. The heat network part of ESDL is
still under heavy development, so there are quite a few TODOs remaining
on our end and that of the specification itself.

ESDL models can be run by inheriting from the ESDLMixin. Do note that
some conventions on pipe naming and interfacing with Delft-FEWS are
assumed by default, but most of these can be overridden if so desired.

[0]: https://energytransition.gitbook.io/esdl/

Closes #21