run_heat_netopt: Fix flow directions detection
We fix the flow direction detection mechanism to handle some corner
cases.
The core idea is: the direction is in one-to-one correspondence with
whether the heat is positive or negative. Moreover, if a pipe can be
disconnected, then we do so when the heat passing through the pipe is
completely consumed by its heat losses.