Examples: Add Westland example case
The Westland case has multiple sources and demands as well as a ring
structure. There is a pump associated to each source (in their
respective cold lines). Due to the ring structure, additional control
valves need to be included in the network to be able to solve for the
head.

Note that the HeatMixin has to find non-obvious flow directions. This
has proven to be complicated when solving for the head in the QTH
problem. Indeed, at the moment, there is no restriction to non-physical
solutions as, for e.g., water flowing in a circle. This would require a
fix to the HeatMixin. As a (temporary) workaround, we limit the pipes
that can have bidirectional flow.