QTH: Minimize head as last priority to catch artificial head loss
The implicit assumption in the model of the QTH Pipe is that additional
head loss can created. This inequality formulation of head loss as a
function of discharge keeps this part of the formulation convex.

In so-called "tree" networks, there is only a single route from source
to demand. Provided there's a controllable valve at every demand, the
additional head loss created by the pipes can be shifted to the valves
at the end without any change in discharge in the system. Note that we
do not actually do this shifting of head losses, as it typically does
not matter for control purposes.

If there is more than one way for water to go from a source to a demand,
this inequality assumption is likely to fail. The QTH optimization might
decide to route all water through the shortest route from source to
demand, whereas the other route gets almost no flow. As long as a
control valve is present in the longer route, this solution can be
realized. If there is no control valve present, the solution cannot be
realized, and the actual flow distribution will potentially differ a lot
from that what we predict.

In such cases, we want to warn the user that our solution not physically
realizable. We also want to convey that there is freedom in the system
to e.g. minimize the overall production, and that a control valve is
needed to leverage this freedom.

To detect these cases, we add an additional `minimize_head_losses`
setting to the `heat_network_options` (enabled by default). When `True`,
the head losses in all pipes _without_ control valves is minimized, as
is the head generated by the pumps. This should move all artificial head
losses to pipes that contain control valves. When this priority is
finished, we check for every pipe without a control valve (and sources
and demands) whether their head loss is as it should be. If it is not,
we log a warning to the user.

To signal whether a pipe contains a control valve (at the start, end or
somewhere in the middle), the `has_control_valve` property is added to
both the Heat and QTH Pipe. To reiterate, in pipes with a control valve
additional head can be created without resulting in a warning.