HeatMixin: Use `keep_soft_constraints` by default
Set goal programming option `keep_soft_constraints` to True in HeatMixin
by default, instead of having to specify it in every case/example.

Note that `keep_soft_constraints` makes sense for the HeatMixin because
it is a fully (mixed-integer) linear model; for a NLP like the QTHMixin
it can more readily lead to convergence issues.