QTHMixin: Add option for quadratic (equality) head loss
The default head loss is the inequality formulation using C*Q^2. This
one uses an estimated velocity (by default 1.0 m/s).

The inequality approach tends to give physically unrealizable results
when water can take multiple routes, and not enough control valves are
present in the system.

Like the `HeadLossOption.LINEAR` linear equality approach added in the
previous commit, we can also offer a quadratic equality formulation.
This tends to not converge at `theta = 0.0`, so we rely on the linear
equality formulation there to give us a good seed for `theta = 1.0`.

Just like for `HeadLossOption.LINEAR`, we use override the head loss
option for pipes with control valves to `HeadLossOption.CQ2_INEQUALITY`.