Test/PyCML: Increase rtol on objective value comparison with Modelica
For some reason, the PyCML model gives a different value for the
objective on the last priority (head minimization) of the QTH problem on
Linux systems. This objective value comparison serves as a sanity check,
so we loosen the relative tolerance to allow these larger (but still
small) differences.

Note that on Windows the results of both models seem to be much closer
for the last priority of the QTH problem (within an rtol of 1e-8). The
objective values of the earlier priorities (and the Heat problem) also
satisfy this tighter tolerance, both on Windows and Linux.