HeatMixin Buffer: Set history when not available
For each buffer, one needs to pass information on how much heat can be
extracted/added to the buffer at t0. This can be done by passing the
history of the `Stored_heat` variable or of `Heat_buffer`.
When history information is not available, we assume that the buffer
will not be used at t0.