Tests: Avoid model/module collisions by using relative imports
We prefer to do relative import in tests to avoid module name
collisions. This is also cleaner than fudging with sys.path to make the
absolute import work. A collision would happen when e.g. multiple tests
would have a file called "model.py", defining the PyCML model. A second
test with `from model import Model` would just use the old (already
loaded) module called `model`. Note that the relative import without
__init__ files works because of PEP 420 (Python >= 3.3).

For debug and developing purposes we also want to run some of the test
files directly. This would set the main module name to "__main__", and
would make the relative import fail. We catch this case, and do an
absolute import instead, as there is no risk of collisions in this case.