Modelica/QTHPort: Fix comment on discharge sign convention
For a QTHTwoPort component, the QTHIn.Q and QTHOut.Q are of equal sign.
The QTHPort's Q variable also does not have a "flow" prefix, which
is congruent with this behavior. The comment stating that "positive flow
is inwards" is likely a remnant of copying from the ChannelFlow
Hydraulic library.