Heat: Add head, discharge and head loss relationship
When we only consider the heat flow rate, we can get solutions that are
hydraulically unrealistic or infeasible. Consider a simple test network
like "DoublePipeEqual", where we would expect flow (and therefore heat
flow rate) to be roughly evenly distributed over the two pipes. However,
before this commit, a solution that would also be perfectly acceptable
was one where a single pipe transfers all heat from the source to the
heat demand, and the other pipe covers only its heat losses. Extending
this behavior to more complex ring networks, there is then no guarantee
that the directions found by looking at the sign of the heat flow rate
match those of the discharge. Indeed, when optimizing a ring network we
would run into infeasibilities when passing the directions to the QTH
problem.

To get a better estimation of the flow directions, and a result that is
more likely to be hydraulically feasible, we add the hydraulics to the
Heat library (its Modelica/PyCML components and the HeatMixin). As the
focus of the HeatMixin is mostly on getting correct flow directions, the
default HeadLossOption is an _equality_ one, to ensure that in e.g.  the
"DoublePipeEqual" test example mentioned above we would end up with a
roughly equal distribution of heat flow rate / discharge (and therefore
pipes having the same flow direction). Any pipes with a control valve
are modeled as an inequality.

The linear head loss relationship uses the `estimated_velocity` by
default for accuracy reasons, and (although unlikely) it can therefore
still result in infeasible solution in theta=0 for the QTH problem
(which uses the CQ2_INEQUALITY option by default). Should this indeed
turn out to be an issue, it is possible to make a linear head loss
relationship based on the _maximum_ velocity. That should guarantee a
feasible solution in the QTH problem, provided the `minimum_velocity` is
set equal in both models. That said, for e.g. a QTHLoopMixin-run it may
be preferable to use the `estimated_velocity`, as that one is also used
by the QTHLoopMixin itself in its theta=0 phase.

Finally, note that head loss on sources is not yet supported.