Buffer: Fix equations when connected pipes have opposite orientation
The equations before this commit implicitly assumed that the hot pipe's
out-port was connected to the buffer's in-port (and similarly that the
buffer's out-port was connected to the cold pipe's in-port). There was
no check on this assumption however.

This commit generalized the equations such that the orientation of the
connecting pipes can be flipped.

Note that we still assume that, throughout the network, the hot and
cold pipes have the same flow direction.