QTH: Set network temperature at theta=0 via bounds
When theta is zero, i.e., the problem is convex, we want to set the
temperature of the cold/hot line to the equal to the designed one.
These values are now set by overwriting the bounds of QTHIn.T/QTHOut.T
of every pipe at theta = 0.0.
Before we used to set the same relationship via path constrains, but
this leads to too many degrees of freedom in ring networks.

Overwriting bounds implies that at theta=0 some constraints would become
trivial (for e.g., heat loss pipe constraints). We skip such constraints.