Examples/Westland: Support changing/shortening optimization horizon
For testing (or closed loop running) it can be useful to temporarily
override the collocation times to only return e.g. the first N
timesteps. This would fail in the Westland example due to the assumption
that `get_timeseries(...).times` and `times()` would return the same
times. A call to `set_timeseries()` requires the times of the passed
Timeseries to be equal to any stored Timeseries, e.g. one retrieved
with a `get_timeseries()` call.