QTHMixin: Improve scaling of pipe heat loss equation
The pipe heat loss equation was not scaled proper. Aside from the large
range in coefficients this may result in for some variables (because of
the `rho * cp`), it also meant that the equation was solve to much
higher accuracy (constraint violation) than necessary. This could lead
to IPOPT struggling to converge, especially if short pipes are present
in the network.