QTHMixin: Improve scaling of Source/Demand Heat constraints
The constraints relating heat of a Source or a Demand to the product of
the discharge and temperature (`rho * cp * Q * dT`) was not taking into
account the typical value for the discharge. This could lead to much
smaller coefficients in constraint than necessary, which in turn could
lead to a larger _range_ of coefficients for a certain variable.
The nominal is set to one by default.

All examples and tests have been updated to also pass the nominal
discharge to Source and Demand components.