Examples/Tests: Fix duplicate/aliased nominals
In some models, a control variable like "Heat_source" was equated to
"Source.Heat_source", and both had a nominal set. The nominals were
different however, which resulted in the _largest_ nominal being chosen,
as is the current Pymoca (0.8.4) behavior. To avoid these mismmatches
and scaling issues that may result from them, it makes sense to only set
the nominal _once_ and let aliasing take care of the rest