QTHMixin: Improve solver stability with new minimum velocity option
If the solver steers towards a solution where the discharge is almost
zero, stability issues can arise. This is even more prevalent when
_seeding_ with a solution where the discharge is almost zero, e.g. when
going from `theta = 0` to `theta = 1`, one goal priority to the next, or
when running in a closed loop that uses seeding.

To avoid these stability issues, a new option is introduced that forces
a minimum velocity in the pipes. This is a value that can be more
reasonably assumed or estimated system-wide than a discharge (as the
latter depends on the diameter of the pipe, of which there may be a wide
range in the system).