Pipe: Remove sign_dT parameter
Cold and hot pipes are assumed to lie close to each other's and thus to
influence the corresponding heat losses. E.g., a cold pipe will receive
some heat from the corresponding hot pipe. Equivalently, the hot pipe
will lose the same amount of heat.
To determine whether a pipe is hot or cold we can rely on the naming
conventions, i.e., a pipe ends with either the '_hot' or the '_cold'
suffix.

Before such conventions were introduced, we used to rely on the
paramter 'sign_dT' to detect whether the pipe was hot or cold. This is
now made obsolete and thus can be safely removed.