Add QTHLoopMixin as fallback for CQ2 equality head loss-discharge
Generally, QTHMixin is used in combination with inequality head-loss
constraints, as those can be formulated in a convex way and thus easier
to solve. For ring networks, or networks with parallel pipes, these
inequality constraints can result in creating artificial head loss (see
commit a85053b8). These solutions are then not hydraulically realizable.

This commit adds an alternative to QTHMixin that solves with
CQ2_EQUALITY head loss constraints, to get solutions that can actually
be realized. To make it easier for the solver, it only solves for two
time steps at a time. Note that this means that for systems with buffers
(or other temporal storage), we need to provide how this temporal
storage should behave (e.g. based on the results of the HeatMixin). See
also the class `BufferTargetDischargeGoal` and the
`buffer_target_discharge_goals` method of QTHLoopMixin.

Also note that, even for very simple networks, there is no guarantee
that QTHLoopMixin will find an equivalent or better solution than using
QTHMixin in combination with CQ2_EQUALITY. It is solely meant as a
fallback to provide _a_ solution of decent quality, without guarantees
on optimality or of a better performance than that of QTHMixin.

The requirement to RTC-Tools is bumped to 2.5.0a4, as this loop
functionality requires the capbility to start the homotopy loop at
theta=1.0 (instead of 0.0).