HeatMixin: Check that Pipe in/out heat flow rate is non-zero
In the HeatMixin there needs to be some heat flow rate into the pipe to
compensate for heat losses. Only in post-processing do we determine if a
pipe was (effectively) disconnected, when one port is zero and the other
matches the heat losses. Doing it like this in post-processing saves us
an integer during the solving.