QTH: Fix H/dH relationship for NO_HEADLOSS option
When the head loss is set to NO_HEADLOSS, we remove any relationship
between H and dH, Q as these will introduce trivial constraints in ring
networks, which might lead to convergency issues of IPOPT.

Note that the H's are completely free when the NO_HEADLOSS option is
chosen. The H's already did not have physical meaning before this
change, as no head relationship was introduced at the demands/sources.