QTH: Add .Q aliased discharge symbol to Source and Demand
The Source and Demand already have discharges on their ports. However,
it is convenient and consistent with e.g. pipes and pumps to also have a
discharge symbol .Q on the components themselves.

Note that this symbol is just an alias of QTHIn.Q, so there is no
calculation overhead.