QTHMixin: Disallow homotopy backtracking
The constraints are setup in such a way that the full homotopy method
can be used.

However, in the linearized version of the problem (i.e., for theta = 0)
the temperatures are assumed to be constant. This means that for
constraints in which temperature changing is fundamental (e.g., heat
losses, node temperature mixing, buffer temperature mixing), there is no
benefit in solving the homotopy process for 0 < theta < 1.

Moreover, the solution at theta = 0 is already a decent seed for the
equations at theta = 1. Thus, we disable the backtracking system of
homotopy mixin, as this will not be beneficial in finding feasible
solutions. This will also, potentially, shorten the computational time
in case of a failed run.