QTHMixin: Fix LINEAR head loss option
The implementation implicitly assumed that the `estimated_velocity` was
always 1.0 m/s. For any other velocity, the correct associated friction
factor was determined, but the linearization of the head loss
relationship still assumed 1.0 m/s.

Because the linear and quadratic relationships start to differ more and
more with this change, they are split into separate branches of the
if-else block.