QTHMixin: Improve scaling of head loss equations/variables
The head loss is typically much smaller than the absolute value of the
upstream and downstream head. To scale the relationship between these
variables better we should:

1. Give the Pipe's .dH symbol a proper nominal

2. Divide the equation relating the heads to the dH symbol with the
   geometric mean of the respective nominals.

We also want to avoid the situation where .dH is sometimes positive-only
(e.g. the inequality options), and sometimes can be both positive and
negative depending on flow direction (LINEAR head loss option).

The formulation that this commit introduces to resolve that is
generalizable to the case where the flow direction is not known a
priori:

3. define the .dH as QTHOut.H - QTHIn.H (i.e. negative when flow is
   positive, and positive when flow is negative)

4. use an additional `.__head_loss` symbol that is equal to the absolute
   value of .dH, but only make this symbol when needed.