Tests/simple_buffer: Make buffer pipes disconnectable
In preparation of adding discharge and head to Heat models, we have a
potential need for this information as part of more than just
post-processing.

Note that the buffer example already has the `disconnectable` property
set on the buffer pipes, as that model is post-processed for directions
(that are sent to the corresponding QTH problem). The simple buffer test
case has no QTH equivalent, so no post-processing was necessary.