Log/print library version when calling `run_heat_network_optimization`
By logging the library version in `run_heat_network_optimization` we
mimic the behavior/signature of RTC-Tools's `run_optimization_problem`.
This includes the default log level of logging.INFO, and being able to
override it by specifying `log_level`. The latter in turn is passed to
`run_optimization_problem` for the Heat and QTH problems.