Heat/Pump: Remove obsolete '.Heat' symbol
The '.Heat' symbol could be used as a proxy for the discharge, such that
bounds on the discharge of the pump could effectively be set. With the
addition of the head and discharge to the Heat models, this ".Heat"
variable for the Pump is no longer necessary.

Note that the symbol was not used in any examples or tests.