QTH/Pump: Remove '.H' symbol
A pump does not really have a "Head" symbol; its upstream head and
downstream head are very different. Just referring to ".H" is not a
conventional way of referring to either of them.

Note that the examples have been adjusted to refer to the _upstream_
head (whereas .H effectively referred to the downstream head).