QTH/Buffer: Add Q_nominal parameter
The Q_hot_pipe and Q_cold_pipe need decent nominals for scaling. We add
a Q_nominal parameter (like many other components) that can be used to
scale these to variables properly. The QTH examples and tests that have
a buffer have been adjusted accordingly to pass this parameter.