QTH/Pipe: Set temperature nominals
The pipe temperature (and node mixing) constraints are particularly
sensitive to scaling. We therefore set some reasonable nominals on the
in- and outgoing temperature of the pipe.

Also clarify that the __init__ method of PyCML's Model should be able to
be called multiple times. For example, for the change in this commit,
"self.temperature" should be set and modified _before_ calling super().