QTH(Loop)Mixin: Fix setting variables to constant via bounds
As the comments near the added lines of code states, it was possible to
end up in a situation where lbx > ubx due to rounding errors in the ulp
(or close to it). These will cause IPOPT to try RELAX_BOUNDS, even
though the upper and lower bound are _really_ close. We avoid this by
making sure that these bounds are exactly equal in case we want them
that way.