QTHLoopMixin: Make 'keep_soft_constraints' the default
We found that issues due to overconstraining the problem are rather
common, especially if some of the additionally added (goal) constraints
are trivial (because we set bounds on some variables at t0). The
downsides of 'keep_soft_constraints' not working well for NLPs (e.g.
not converging) generally are more applicable to larger optimization
problems. So for the QTHLoopMixin the benefits of using the option
'keep_soft_constraints' outweigh the drawbacks.

We also bump the tolerance on the tests a little. With the changes of
this commit they were violated ever so slightly with the default rtol of
1e-7 and atol of zero.