QTHMixin: Remove dtemp option
At every demand we impose a temperature jump correspoding to dT. We used
to set this dT as a heat_network_option. This is no longer the case, as
dT should always be the difference in temperature between the default
supply and return line, which is part of every component's parameters.
We should therefore just use these parameters directly instead.