Fix MRO/inheritence conflicts
Due to the way we inherited, it was possible for MRO conflicts to arise.
This is particularly related to the base classes of the ESDLMixin. We
now make ESDLMixin explicitly depend on ModelicaComponentTypeMixin (as
it does not provide its own implementation Topology yet), and adjust the
test models accordingly.

We also need to be sure that ESDLMixin defines `def times` (a
CollocatedIntegratedOptimizationProblem). Although IOMixin does define
this method, it is optional, i.e. RTC-Tools does not require users using
IOMixin to _also_ use CollocatedIntegratedOptimizationProblem. ESDLMixin
therefore needs to explicitly inherit from the latter.