Docs: Add background theory, approach, and Python API documentation
Note that we do not (yet) include documentation for:

- The examples
- The Modelica/PyCML/component library

The scaffolding for the examples is part of this commit, such that it
can easily be added later.

Whether documentation of the component library will be added is unclear
at the time of this commit. For now, the recommendation is to use ESDL
models, and refer to the respective documentation.