Make Source component also always act as a Pump
To align ourselves with how e.g. CHESS models systems, i.e. not
explicitly modeling pumps, we incorporate the head production
capabilities of a Pump into Source components.

Note that the Pump model is very basic at the moment, and indeed only
considers production of head (with a possible maximum). There is no
intricate pump-curve information yet. Should that information become
available and necessary to the optimization, it _would_ make sense for
users to explicitly model Pump components, and we would have to take
this additional detail into account. At that point in time, we could
(manually or automatically) set the production of head by Sources to
zero, e.g. by forcing .dH to be zero.

One test case fails to converge if we leave the dH of the source free,
due to the large degree of freedom this leaves the solver. Giving the dH
of the source any reasonable upper bound solves this, but for simplicity
we explicitly set the dH of the source to zero.