QTH/Buffer: Fix convergence issue by reformulating pipe heat losses
The heat losses in the buffer are linearly dependent on the volume.
Specifically the heat losses only depend on the curved surface of the
buffer, not on the top/bottom. The previous formulation, that included
the bottom, lead to convergence issues when the volume of the buffer
was close to its lower bound.
Note that this new formulation matches the one of the HeatMixin.