QTH/Buffer: Set non-zero volume lower bounds
We restrict the volume of each tank to always be a fraction of the total
maximum volume, by default 5%. This can be set using the
`min_fraction_tank_volume` parameter.

We make this assumption for two reasons:
* avoid weird temperature mixing effects due to the negligible volumes;
* include a reasonable operational assumption in the system.

The user can set the initial volume of the hot tank of the buffer.
If not specified, the initial volume is the minimum possible.

For what concerns the ESDL the capacity is assumed to be the relative
capacity of the buffer.