Heat/Buffer: Set non-zero stored heat lower bounds
Similarly as QTH - commit 0b34a6c6, we restrict the tank capacity
such that the tanks are always at least a fraction of the total
volume, by default 5%. This can be set using the parameter
`min_fraction_tank_volume`.

The user can set the initial stored heat of the buffer either by
providing the initial stored heat itself or the initial volume of the
hot tank. At most one of these values can be set. If neither is
provided, the minimum possible stored heat is taken as the initial
value. The initial stored heat is set by overwriting the bounds at t0.

In the ESDL, the buffer capacity is assumed to be the relative total
heat that can be stored in the buffer.