HeatMixin: Fix head loss symbol used for cold pipes
Cold and hot pipes each have their own head loss symbol. If we use the
same symbol, we get warnings for cold pipes when we enable the
`minimize_head_loss` option.

Note that in many cases the head losses of a hot and cold pipe are
almost equal, so results were unaffected. Generally one does not run the
`minimize_head_loss` option in HeatMixin, and therefore would not
encounter the warnings either.

For some reason, CBC struggles with finding a solution in priority 2 of
the heat problem of Westland with this change implemented. Passing the
solution of the previous priority fixes this. For now, we ascribe this
behavior to CBC being weird, but it may require a more general fix in
the future.