Source: Add 'price' parameter
This can be used to characterize the costs of a Source. In turn, this
parameter can then be used in minimization objectives, to prefer one
source over another.

Note that a small bump in tolerance was required for a single test.
Apparently there is an ever so slight difference in the DAE between the
PyCML and Modelica versions.