EsdlHeatBuffer: Set min_fract_tank_volume to zero
If the min_fract_tank_volume is non-zero, the buffer will have to be
charged in the first timestep (due to heat losses). This should not be
the case and can be avoided by setting min_fract_tank_volume to zero.
Note that the same change is not relevant for QTH.