ESDL: Skip handling of Sensor-type assets
Sensors only have a single port, so we need to allow those types of
components in esdl_mixin.py. The types of the Asset class have been
changed accordingly (in_port and out_port now being Optional).

In the handling of components, we want to _skip_ the sensor in both the
Heat and QTH models. We therefore add a default "skip" conversion
method. Lastly, we want to avoid adding connections to ports of
components that do not exist (or we would get a KeyError).

We implicitly always assumed the number of connected components to be
exactly 1 for all but the nodes. We now make this check explicit.