Add support for control valves and check valves
Check valves are valves that enforce one-way flow (Q >= 0). We implement
an idealized version of a check valve that has the following properties:

- dH = 0 when Q > 0 (open)
- dH > 0 when Q = 0 (closed)

In other words, if the pressure downstream of the check valve is higher
than upstream, the check valve is closed. There is no minimum pressure
drop required to open the check valve, and neither do we take into
account the head loss that the check valve itself causese.

Control valves are basically zero-length pipes. They are implemented
according to the following properties:

- dH > 0 when Q < 0
- dH < 0 when Q > 0

The exact value of dH is free for any given discharge, as long as the
sign matches these criteria. We neglect any head loss that the control
has in fully opened position (dH = 0 in our implemention).

As is clear from the above two sets of properties, the implementation of
check and control valves involves booleans in the HeatMixin. Similarly,
it requires that the open/closed status of a check valve and the flow
direction of a control valve is fixed (not free to choose), just like
the flow direction of a pipe is.

To that end, two new enums are defined that handle the status of a check
valve (`CheckValveStatus`) and the flow direction of a control valve
(`ControlValveDirection`). These enum values are passed with the
`flow_directions` keyword argument to the QTHMixin. Similarly, the
`heat_network_pipe_flow_directions` property of QTHMixin has been
renamed ("pipe" is removed), and returns the flow directions of check
valves and control valves as well.

In the HeatMixin the booleans are used with big-M formulations. For that
it is needed to have an upper bound of both the discharge and the head
loss.  E.g. when a check valve is closed, we need to know how high the
dH can get. For the discharge we use the maximum discharge of the
connected pipes. For the head loss we add a helper routine to sum up all
the contributions to head loss in the system, which currently are pipes
and demands. For this calculation to hold, the assumption is that pumps
will never generate _more_ head than necessary to accomplish a certain
discharge. In other words, for a given discharge, we assume that pumps
are able to deliver it at the lowest possible head. Note that for e.g.
fixed-speed pumps this assumption does not hold.